CacheThread moved to FATHandler cause it needs SCSI access
